+++ cytoscape/branches/cytoscape_28_custom_graphics/src/cytoscape/visual/ui/editors/continuous/C2CMappingEditor.java
@@ -240,7 +240,7 @@ public class C2CMappingEditor extends ContinuousMappingEditorPanel {
 		for (ContinuousMappingPoint point : allPoints) {
 			bound = point.getRange();
 
-			fraction = ((Number) ((point.getValue() - minValue) / actualRange)).floatValue() * 100;
+			fraction = ((Number) ((point.getValue().doubleValue() - minValue) / actualRange)).floatValue() * 100;
 			slider.getModel().addThumb(fraction, ((Number) bound.equalValue).floatValue());
 		}
